backend/worker/requirements.txt
redis
pymongo
requests
ffmpeg-python
opencv-python-headless
numpy
